Title: 배열에서 원소의 특징, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240406/배열에서 원소의 특징/characteristics-of-elements-in-an-array.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main() {
     // 여기에 코드를 작성해주세요.
-    double arr[10];
+    int arr[10];
 
     for(int i = 0; i < 10; i++){
         cin >> arr[i];
